+++ plugins/CoreHome/CoreHome.php
@@ -380,5 +380,11 @@ class CoreHome extends \Piwik\Plugin
         $translationKeys[] = 'General_DoubleClickToChangePeriod';
         $translationKeys[] = 'General_Apply';
         $translationKeys[] = 'General_Period';
+        $translationKeys[] = 'CoreHome_ShortcutZenMode';
+        $translationKeys[] = 'CoreHome_ShortcutSegmentSelector';
+        $translationKeys[] = 'CoreHome_ShortcutWebsiteSelector';
+        $translationKeys[] = 'CoreHome_ShortcutCalendar';
+        $translationKeys[] = 'CoreHome_ShortcutSearch';
+        $translationKeys[] = 'CoreHome_ShortcutHelp';
     }
 }