+++ tests/UI/specs/UIIntegration_spec.js
@@ -81,6 +81,14 @@ describe("UIIntegrationTest", function () { // TODO: Rename to Piwik?
         }, done);
     });
 
+    // shortcuts help
+    it("should show shortcut help", function (done) {
+        expect.screenshot("shortcuts").to.be.captureSelector('.modal.open', function (page) {
+            page.load("?" + urlBase + "#?" + generalParams + "&category=Dashboard_Dashboard&subcategory=1");
+            page.sendKeys('body', '?', 100);
+        }, done);
+    });
+
     // visitors pages
     it('should load visitors > overview page correctly', function (done) {
         expect.screenshot("visitors_overview").to.be.captureSelector('.pageWrap', function (page) {